// types/user-register.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface UserRegisterSchema {
  name: string;
  address: string;
  email: string;
  password: string;
  creditcard_type: string;
  creditcard_name: string;
  creditcard_number: string;
  creditcard_expired: string;
  creditcard_cvv: string;
  [k: string]: unknown;
}